DNS iterative Resolver
Developed DNS Resolver in python which queries the appropriate chain of
DNS authoritative servers to resolve a domain name for any requested
record type, include A records, AAAA records, MX records, TXT records,
etc.
The resolver should store answers in a cache, and retrieve answers from
cache if present. Each step of the resolution should be timed so that we
can get the latency for each step of the resolution as well as overall
latency to get the final answer.
Technologies: Python ,Unix System